Rubocop errors fix

+++ app/helpers/application_helper.rb
@@ -1,7 +1,5 @@
 module ApplicationHelper
-    def links_display
-        if user_signed_in?
-          link_to('Logout', destroy_user_session_path, method: :delete)
-        end
-    end
+  def links_display
+    link_to('Logout', destroy_user_session_path, method: :delete) if user_signed_in?
+  end
 end
